backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
pinecone-client==3.1.0
python-dotenv==1.0.0
numpy==1.26.3
pydantic==2.6.1
python-multipart==0.0.6
gunicorn==21.2.0